Rebuild missing zip files
If a Lambda function is manually deleted outside of Terraform, or in other weird cases that we don't understand, Terraform might try to upload a zip file that doesn't exist on disk.

Zip files are now kept for 7 days and if they don't exist, but might need to be uploaded, then a new data source will rebuild them.

This can result in zip files being built by a data source, where the built output won't be visible anywhere. This should only happen when the lambda function source code hasn't changed and it won't be uploaded (so you don't really care about the build output) or in the edge cases where it is trying to upload a file and it doesn't exist (not ideal to lose the build output, but it is better than getting an error). Outside of those 2 cases, the zip will be built by the null_resource and the output will be visible as expected.