Fix Docker permission issues
When I tried this, it was creating the Lambda build files as the root user and then it failed trying to clean them up. This now does everything including setting ownership inside Docker.

I also removed the default values because I'd rather it break then default to something that may not be intended.