Upgrade module to Terraform 0.12.x
I could not find a way to support 0.11.x and 0.12.x at the same time, so this goes all-in on 0.12.x features and cleans up a lot of code in the process.

terraform-docs does not support 0.12.x yet so I've removed the Makefile and manually wrote the README.

I have tried all tests in the test directory and they all worked.